requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
idna==2.10
PyGeodesy==20.9.29
python-aqi==0.6.1
requests==2.24.0
urllib3==1.25.10
